+++ packages/use-async/src/cache.js
@@ -1,11 +1,11 @@
-import { useEffect } from 'react';
+import get from '@postinumero/map-get-with-default';
 import stringify from 'fast-json-stable-stringify';
 import memoize from 'memoizee';
-import get from '@postinumero/map-get-with-default';
+import { createContext, useContext, useEffect } from 'react';
 
-const cache = new WeakMap();
+const globalCache = new WeakMap();
 
-export function getItem(func, config, args) {
+export function getItem(func, config, args, cache = globalCache) {
   return get(
     get(cache, func, () => new WeakMap()),
     config,
@@ -23,15 +23,18 @@ export function getItem(func, config, args) {
   )(...args);
 }
 
-function removeItem(func, config, args) {
+function removeItem(func, config, args, cache = globalCache) {
   cache
     .get(func)
     .get(config)
     .delete(...args);
 }
 
+export const CacheContext = createContext(globalCache);
+
 export function useItem(...args) {
-  const memoized = getItem(...args);
+  const cache = useContext(CacheContext);
+  const memoized = getItem(...args, cache);
   useEffect(() => {
     const { updaters, cleanupTimeout } = memoized;
     if (cleanupTimeout) {
@@ -41,7 +44,7 @@ export function useItem(...args) {
       memoized.cleanupTimeout = setTimeout(() => {
         if (!updaters.size) {
           memoized.cancel?.();
-          removeItem(...args);
+          removeItem(...args, cache);
         }
       }, 0);
     };