+++ packages/use-async/src/loading-state/create.js
@@ -1,4 +1,4 @@
-import { constructor } from '../create';
+import { constructor } from '../create.js';
 import recall from './recall.js';
 import useAsync from './useAsync.js';
 import useAsyncSafe from './useAsyncSafe.js';